유치원 소개 페이지와 유치원 관련 관리자 페이지 구조 변경
